requirements.txt
fastapi
uvicorn
caspyorm
pydantic
python-dotenv 